--- 01字符串转换成二维码.py
@@ -1,15 +1,0 @@
-from PIL import Image
-MAX = 25    #宽高 625个01字符串，所以宽和高是25
-pic = Image.new("RGB",(MAX, MAX))
-#若是(255,255,255)、(0,0,0)需要先转换01 
-str = "1111111000100001101111111100000101110010110100000110111010100000000010111011011101001000000001011101101110101110110100101110110000010101011011010000011111111010101010101111111000000001011101110000000011010011000001010011101101111010101001000011100000000000101000000001001001101000100111001111011100111100001110111110001100101000110011100001010100011010001111010110000010100010110000011011101100100001110011100100001011111110100000000110101001000111101111111011100001101011011100000100001100110001111010111010001101001111100001011101011000111010011100101110100100111011011000110000010110001101000110001111111011010110111011011"
-i=0
-for y in range (0,MAX):
-    for x in range (0,MAX):
-        if(str[i] == '1'):
-            pic.putpixel([x,y],(0, 0, 0))
-        else:
-            pic.putpixel([x,y],(255,255,255))
-        i = i+1
-pic.show()
-pic.save("./flag.png")
