+++ 14.盲注判断密码.py
